Repackage properties, create ObjectProperty interface and implementation
