--- modules/api/src/main/java/net/countercraft/movecraft/craft/type/DefaultProvider.java
@@ -1,6 +1,0 @@
-package net.countercraft.movecraft.craft.type;
-
-@FunctionalInterface
-public interface DefaultProvider<Type> {
-    Type process(CraftType type);
-}
