+++ modules/api/src/main/java/net/countercraft/movecraft/craft/type/property/DefaultProvider.java
@@ -1,0 +1,8 @@
+package net.countercraft.movecraft.craft.type.property;
+
+import net.countercraft.movecraft.craft.type.CraftType;
+
+@FunctionalInterface
+public interface DefaultProvider<Type> {
+    Type process(CraftType type);
+}
