+++ modules/api/src/main/java/net/countercraft/movecraft/craft/type/property/IntegerProperty.java
@@ -1,29 +1,31 @@
-package net.countercraft.movecraft.craft.type;
+package net.countercraft.movecraft.craft.type.property;
 
+import net.countercraft.movecraft.craft.type.CraftType;
+import net.countercraft.movecraft.craft.type.TypeData;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class FloatProperty {
+public class IntegerProperty {
     private final String key;
-    private final DefaultProvider<Float> defaultProvider;
+    private final DefaultProvider<Integer> defaultProvider;
 
     /**
-     * Construct a FloatProperty
+     * Construct an IntegerProperty
      *
      * @param key the key for this property
      */
-    public FloatProperty(@NotNull String key) {
+    public IntegerProperty(@NotNull String key) {
         this.key = key;
         this.defaultProvider = null;
     }
 
     /**
-     * Construct a FloatProperty
+     * Construct an IntegerProperty
      *
      * @param key the key for this property
      * @param defaultProvider the provider for the default value of this property
      */
-    public FloatProperty(@NotNull String key, @NotNull DefaultProvider<Float> defaultProvider) {
+    public IntegerProperty(@NotNull String key, @NotNull DefaultProvider<Integer> defaultProvider) {
         this.key = key;
         this.defaultProvider = defaultProvider;
     }
@@ -36,9 +38,9 @@ public class FloatProperty {
      * @return the value
      */
     @Nullable
-    public Float load(@NotNull TypeData data, @NotNull CraftType type) {
+    public Integer load(@NotNull TypeData data, @NotNull CraftType type) {
         try {
-            return (float) data.getDouble(key);
+            return data.getInt(key);
         }
         catch (IllegalArgumentException e) {
             if(defaultProvider == null)
