+++ modules/api/src/main/java/net/countercraft/movecraft/craft/type/property/ObjectProperty.java
@@ -1,0 +1,14 @@
+package net.countercraft.movecraft.craft.type.property;
+
+import net.countercraft.movecraft.craft.type.CraftType;
+import net.countercraft.movecraft.craft.type.TypeData;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public interface ObjectProperty {
+    @Nullable
+    Object load(@NotNull TypeData data, @NotNull CraftType type);
+
+    @NotNull
+    String getKey();
+}
